T5_Model/requirements.txt
sentencepiece >= 0.1.95
transformers >=4.6.1
datasets >=1.6.2
torchvision
tensorboard
